Introduce PushNotificationsUtility.h/cpp
